Move open telemetry config to own package and fix build
