+++ src/cmd/services/m3dbnode/config/config.go
@@ -42,7 +42,7 @@ import (
 	"github.com/m3db/m3/src/x/opentracing"
 
 	"github.com/m3dbx/vellum/regexp"
-	"go.etcd.io/etcd/pkg/transport"
+	"go.etcd.io/etcd/client/pkg/v3/transport"
 	"go.etcd.io/etcd/pkg/types"
 	"go.etcd.io/etcd/server/v3/embed"
 )
